move models to models module
