+++ docker/model_download.py
@@ -2,8 +2,7 @@ import os
 
 from transformers import AutoConfig, AutoTokenizer
 
-from cnlpt.CnlpModelForClassification import CnlpModelForClassification
-from cnlpt.HierarchicalTransformer import HierarchicalModel
+from cnlpt.models import CnlpModelForClassification, HierarchicalModel
 from cnlpt.train_system import is_hub_model
