+++ src/cnlpt/models/baseline/__init__.py
@@ -1,0 +1,4 @@
+from .cnn import CnnSentenceClassifier
+from .lstm import LstmSentenceClassifier
+
+__all__ = ["CnnSentenceClassifier", "LstmSentenceClassifier"]
