+++ src/cnlpt/train_system.py
@@ -42,14 +42,13 @@ from transformers import (
 from transformers.file_utils import CONFIG_NAME
 from transformers.training_args import IntervalStrategy
 
-from .BaselineModels import CnnSentenceClassifier, LstmSentenceClassifier
 from .cnlp_args import CnlpTrainingArguments, ModelArguments
 from .cnlp_data import ClinicalNlpDataset, DataTrainingArguments, get_dataset_segment
 from .cnlp_metrics import cnlp_compute_metrics
 from .cnlp_predict import process_prediction, restructure_prediction, structure_labels
 from .cnlp_processors import relex, tagging
-from .CnlpModelForClassification import CnlpConfig, CnlpModelForClassification
-from .HierarchicalTransformer import HierarchicalModel
+from .models import CnlpConfig, CnlpModelForClassification, HierarchicalModel
+from .models.baseline import CnnSentenceClassifier, LstmSentenceClassifier
 
 sys.path.append(os.path.join(os.getcwd()))
